Refactor: Reorder first loop body of list_ticket()
Before functionizing it.

This adds nesting and creates some overly long lines, but I'm not
concerned about it, since this will be addressed in an upcoming commit.

The idea here is that there are certain strings (timeout_str, until_str,
and pending_str) that we use only if is_owned(tk) is true. Previously,
we built those strings in advance, even if is_owned(tk) would turn out
to be false. Presumably, this was to reduce nesting. However, I found it
confusing. We were building strings well before we appended them to the
buffer (s), and they appeared out of order.

Now, strings appear in the order in which they're appended to the
buffer, scopes are kept as narrow as possible, and we drop the
pending_str variable.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>