Doc: Clarify non-obvious logic of get_other_site()
I was confused about why the meaning of *node == NULL changed between
the inside and outside of the loop, since the only thing that could
change it is the loop body itself. So I'm adding a couple of comments to
clarify.

Support for OTHER_SITE began with commit a9d48c98. That commit also
updated the docs to say "Obviously, in that case, the booth
configuration must have exactly two sites defined."

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>