Refactor: Drop enum match_type from transport.c
And return void from find_address(). Nothing used the return value.

The only remaining way we used enum match_type was for the did_match
variable, which worked as follows.
* If we found an exact match, we set did_match to EXACT_MATCH and broke
  out of the loop.
* If we found an inexact match, we checked whether did_match was less
  than EXACT_MATCH. If so, we set did_match to FUZZY_MATCH and continued
  to the next iteration.

However, it was impossible for did_match to be set to EXACT_MATCH
without breaking out of the loop. So the check mentioned above always
evaluated to true.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>