Refactor: Move node_to_addr_pointer() to transport.c
Since it's used only there. Also rename it to get_site_addr() and have
it take a const argument and return a const pointer.

Store the return value in an unsigned char pointer. We're comparing
element by element against ipaddr, which is an unsigned char array. The
types should be the same.

Rename the node variable to site and the n_a variable to site_addr.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>