Refactor: Rename struct booth_config:transport to udp
I don't see a need for modularity unless/until we actually want to make
this configurable with more than one option.

As it stands right now, for any given struct booth_config, we have
exactly two sets of transport functions (TCP and UDP), and we use each
of them in specific, pre-defined contexts.

An upcoming commit will add a tcp field that corresponds to the udp
field.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>