Refactor: Drop BOOTH_OPT_AUTH
This was added by the following commits:
* 685a2789: Added the iv field, which later became the opts field.
* d17c6776: Set the iv field to htonl(1) if is_auth_req().
* 9ca5e67d: Created BOOTH_OPT_AUTH as an enum alias for 1.

However, that was 11 years ago in 2014. Nothing checked the iv/opts
field back then, and as far as I can tell, nothing has ever checked it.
We've been setting this bit just in case something wants to check it in
the future in a mixed-version booth cluster.

There are still no plans to add anything that would check this field.
The only reason I'm keeping the opts field is that shifting the values
in struct boothc_header would break parsing of messages from older
clients.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>