Refactor: Defunctionize init_header_bare()
It serves no apparent purpose. It makes init_header() shorter, but it
doesn't aid in understanding because this isn't a logical division.
init_header_bare() just does the parts of initialization that depend
only on conf and h.

It was added by commit ff0bd267 in 2013.

Also, replace tabs with spaces in init_header(), and add more
assertions.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>