Fix: Parse {site,arbitrator}-{user-group} from config file correctly
This feature has apparently been broken since it was added by commit
e5cfc643 in 2014. optarg is a global variable that is set by getopt() in
read_arguments(). When we're parsing the configuration file, we've
already finished parsing the command line arguments and returned from
read_arguments(). optarg is set to whatever was the value of the last
command line option. That value is meaningless in the current context.

We should instead be using the value from the configuration file line
that we're parsing.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>