Feature: Support quotes in attr-prereq
An attr-prereq directive in the booth configuration file can now contain
quoted values. For example:

    attr-prereq = auto test_attr eq "this is my value"

now assigns "this is my value" (without the double quotes) as the
attribute's value.

Technically, this support extends to every token in the attr-prereq
directive. However, this is just a quirk of the implementation, rather
than the intent. For example,

    attr-prereq = 'auto' "test_attr" 'eq' "this is my value"

is equivalent to the directive above.

The main impetus for this change was to get rid of the strtok() calls,
which are harder to read than g_shell_parse_argv().

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>