Refactor: Move do_parse_weights() body into parse_weights()
Note: the weights option gets deprecated a couple of commits from now,
and this code gets removed. I noticed later that we could do that. I've
left this commit and the next one just in case we need to revert the
removal for some reason.

* Replace tabs with spaces.
* Replace -1 (error) with false and i (success) with true.
* Replace input with value and weights with (*ticket)->weight so that
  the code will compile.

Otherwise, no code changes in this commit.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>