Refactor: Drop action argument from booth__read_config()
It's used only by one parser function, and we can get it from cl.type.

Eventually it would be nice to drop cl as a global struct, and at that
point we'll need to pass in the argument again (even if only to set it
as a field of struct booth_config). It's easy to add it back when we
need it.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>