Refactor: Pass a parser_context struct to parser functions
Changing the definition of a struct is easier than changing argument
lists for a bunch of functions, which is something we've been doing a
fair amount of in previous commits.

It also makes more sense to store the current ticket in a parser context
struct. Previously, within the parser functions, it wasn't totally clear
that the ticket argument referred to the ticket we were currently
parsing (which was presumed to be set by a previous parser function).

I've left some overly long lines, because I expect the error logging
scheme to change soon, and it's not worth adjusting everything.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>